refactor : try to streamline token loading into component
